Migrates Main behavior up to index. Corrects test typing.
