+++ tests/pages/index.test.tsx
@@ -1,7 +1,8 @@
 import renderer from 'react-test-renderer'
 import { render, screen } from '@testing-library/react'
 import Index from '../../pages/index'
-import getScenarioContent from '../../utils/getScenarioContent'
+import getScenarioContent, { ScenarioType } from '../../utils/getScenarioContent'
+import apiGatewayStub from '../../utils/apiGatewayStub'
 import { ScenarioContent } from '../../types/common'
 
 import { useRouter } from 'next/router'
@@ -14,8 +15,7 @@ jest.mock('next/router', () => ({
 let scenarioContent: ScenarioContent
 
 beforeAll(() => {
-  const pendingDeterminationScenario = { pendingDetermination: ['temporary text'] }
-  scenarioContent = getScenarioContent(pendingDeterminationScenario)
+  scenarioContent = getScenarioContent(apiGatewayStub(ScenarioType.PendingDetermination))
 })
 
 describe('Exemplar react-test-renderer Snapshot test', () => {
@@ -30,9 +30,31 @@ describe('Exemplar react-test-renderer Snapshot test', () => {
   })
 })
 
-describe('Example react testing-library Test', () => {
-  it('has our placeholder app', () => {
-    render(<Index loading={false} scenarioContent={scenarioContent} />)
+describe('Main component shows loading', () => {
+  it('has our titles but content is shimmer', () => {
+    const mockRouter = {
+      locale: 'en',
+    }
+    ;(useRouter as jest.Mock).mockReturnValue(mockRouter)
+
+    render(<Index loading scenarioContent={scenarioContent} />)
     expect(screen.queryByText('Claim Tracker')).toBeInTheDocument()
+    expect(screen.queryByText('Claim Status')).toBeInTheDocument()
+    expect(screen.queryByText('Next Steps')).toBeInTheDocument()
+    expect(screen.queryByText('Benefit Year')).not.toBeInTheDocument()
+    expect(screen.queryByText('Claim Status')).toBeInTheDocument()
+    expect(screen.queryByText('Your payments')).not.toBeInTheDocument()
+  })
+})
+
+describe('Main component shows the timeout', () => {
+  it('has our timeout modal', () => {
+    const mockRouter = {
+      locale: 'en',
+    }
+    ;(useRouter as jest.Mock).mockReturnValue(mockRouter)
+
+    render(<Index timedOut loading={false} scenarioContent={scenarioContent} />)
+    expect(screen.queryByText('Your Session Will End Soon')).toBeInTheDocument()
   })
 })
